added helper_cis for model MgH+ calculation with PF Hamiltonian
helper_cis.py returns electronic energy of first excited-state along with the total dipole moment associated with the ground- and first-excited state, and the transition dipole moment between the two.  The former 2 quantities are used in the dipole self energy and the latter quantity is used in the bilinear coupling.