made some progress!
Found a big source of error in the CQED-CIS equations... some quantities, like the eris, are slices of the integrals (oovv and ovov), so the virtual indices for these arrays start from zero.  However, the quadrupole and dipole arrays are not slices, so the virtual indices need to start from ndocc!  This has been fixed, and a first pass at the H_ias,jbt Hamiltonian in the cqed_rhf basis has been made.  Will look into this further in the next few days and hopefully submit a pull request soon!